Se refactoriza arquitectura de rutas
