+++ Servidor/.gitignore
@@ -3,7 +3,7 @@ node_modules
 package-lock.json
 
 # Enviroments
-.env
+# .env
 
 # MacOS
 .DS_Store
